Extract DockerContainer into its own file
